refactor(coins): rename app module's package `presentation` -> `ui`
