+++ app/src/main/java/com/hxl/cryptonumismatist/ui/fragments/BaseFragment.java
@@ -1,4 +1,4 @@
-package com.hxl.cryptonumismatist.presentation.fragments;
+package com.hxl.cryptonumismatist.ui.fragments;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
