+++ app/src/main/java/com/hxl/cryptonumismatist/ui/fragments/welcome/WelcomeFragment.java
@@ -1,4 +1,4 @@
-package com.hxl.cryptonumismatist.presentation.fragments.welcome;
+package com.hxl.cryptonumismatist.ui.fragments.welcome;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -13,7 +13,7 @@ import androidx.viewpager2.widget.ViewPager2;
 
 import com.hxl.cryptonumismatist.R;
 import com.hxl.cryptonumismatist.databinding.FragmentWelcomeBinding;
-import com.hxl.cryptonumismatist.presentation.fragments.BaseFragment;
+import com.hxl.cryptonumismatist.ui.fragments.BaseFragment;
 import com.hxl.presentation.viewmodels.WelcomeViewModel;
 
 import dagger.hilt.android.AndroidEntryPoint;
