+++ app/src/main/res/layout/fragment_coin_details.xml
@@ -5,5 +5,5 @@
     <FrameLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.fragments.coins.details.CoinDetailsFragment" />
+        tools:context=".ui.fragments.coins.details.CoinDetailsFragment" />
 </layout>
